Go: remove uses of the lock file
This is now handled by qcow-tool itself for `compact`.  It was
actually not needed for `check` (see the pinata PR): it is read-only
and designed to work on a live disk.

See https://github.com/mirage/ocaml-qcow/pull/106
and https://github.com/docker/docker-diagnose/pull/33
and https://github.com/docker/pinata/pull/8430.

Signed-off-by: Akim Demaille <akim.demaille@docker.com>